+++ starlite/logging/picologging.py
@@ -1,6 +1,4 @@
 import atexit
-from io import StringIO
-from logging import StreamHandler
 from queue import Queue
 from typing import Any, List, Optional
 
@@ -8,6 +6,7 @@ from starlite.exceptions import MissingDependencyException
 from starlite.logging.utils import resolve_handlers
 
 try:
+    from picologging import StreamHandler
     from picologging.handlers import QueueHandler, QueueListener
 except ImportError as e:
     raise MissingDependencyException("picologging is not installed") from e
@@ -28,7 +27,7 @@ class QueueListenerHandler(QueueHandler):  # type: ignore[misc]
         if handlers:
             handlers = resolve_handlers(handlers)
         else:
-            handlers = [StreamHandler(StringIO())]
+            handlers = [StreamHandler()]
         self.listener = QueueListener(self.queue, *handlers)
         self.listener.start()