+++ starlite/template/base.py
@@ -1,18 +1,24 @@
-from typing import TYPE_CHECKING, Any, List, Mapping, TypeVar, Union, cast
+from typing import TYPE_CHECKING, Any, Callable, Dict, List, TypeVar, Union
 
 from pydantic import DirectoryPath, validate_arguments
-from typing_extensions import Protocol, runtime_checkable
+from typing_extensions import Protocol, TypedDict, runtime_checkable
 
 if TYPE_CHECKING:
     from starlite import Request
 
 
-def url_for(context: Mapping[str, Any], route_name: str, **path_parameters: Any) -> str:
+class TemplateContext(TypedDict):
+    request: "Request[Any, Any]"
+    csrf_input: str
+
+
+def url_for(context: TemplateContext, route_name: str, **path_parameters: Any) -> str:
     """Wrapper for [route_reverse][starlite.app.route_reverse] to be used in
     templates.
 
     Args:
-        name: A route handler unique name.
+        context: The template context.
+        route_name: The name of the route handler.
         **path_parameters: Actual values for path parameters in the route.
 
     Raises:
@@ -21,8 +27,24 @@ def url_for(context: Mapping[str, Any], route_name: str, **path_parameters: Any)
     Returns:
         A fully formatted url path.
     """
-    request = cast("Request", context.get("request"))
-    return request.app.route_reverse(route_name, **path_parameters)
+    return context["request"].app.route_reverse(route_name, **path_parameters)
+
+
+def csrf_token(context: TemplateContext) -> str:
+    """Sets a CSRF token on the template.
+
+    Notes:
+        - to use this function make sure to pass an instance of [CSRFConfig][starlite.config.csrf_config.CSRFConfig] to
+        the [Starlite][starlite.app.Starlite] constructor.
+
+    Args:
+        context: The template context.
+
+
+    Returns:
+        A CSRF token if the app level `csrf_config` is set, otherwise an empty string.
+    """
+    return context["request"].scope.get("_csrf_token", "")  # type: ignore
 
 
 class TemplateProtocol(Protocol):  # pragma: no cover
@@ -71,3 +93,14 @@ class TemplateEngineProtocol(Protocol[T_co]):  # pragma: no cover
             [TemplateNotFoundException][starlite.exceptions.TemplateNotFoundException]: if no template is found.
         """
         ...
+
+    def register_template_callable(self, key: str, template_callable: Callable[[Dict[str, Any]], Any]) -> None:
+        """Registers a callable on the template engine.
+
+        Args:
+            key: The callable key, i.e. the value to use inside the template to call the callable.
+            template_callable: A callable to register.
+
+        Returns:
+            None
+        """