+++ tests/dto_factory/test_dto_factory_model_conversion.py
@@ -5,13 +5,22 @@ from typing import Any, Callable, Dict, List, Type, cast
 
 import pytest
 from pydantic_factories import ModelFactory
+from typing_extensions import is_typeddict
 
 from starlite import DTOFactory, ImproperlyConfiguredException
 from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
 from starlite.plugins.tortoise_orm import TortoiseORMPlugin
-from tests import Person, Species, VanillaDataClassPerson
+from tests import Person, Species, TypedDictPerson, VanillaDataClassPerson
 from tests.plugins.sql_alchemy_plugin import Pet
 from tests.plugins.tortoise_orm import Tournament
+
+
+def _get_attribute_value(model_instance: Any, key: str) -> Any:
+    """Utility to support getting values from a class instance, or dict."""
+    try:
+        return model_instance.__getattribute__(key)
+    except AttributeError:
+        return model_instance[key]
 
 
 @pytest.mark.parametrize(
@@ -19,6 +28,7 @@ from tests.plugins.tortoise_orm import Tournament
     [
         [Person, [], {"complex": "ultra"}, []],
         [VanillaDataClassPerson, [], {"complex": "ultra"}, []],
+        [TypedDictPerson, [], {"complex": "ultra"}, []],
         [Pet, ["age"], {"species": "kind"}, [SQLAlchemyPlugin()]],
     ],
 )
@@ -34,10 +44,12 @@ def test_conversion_to_model_instance(model: Any, exclude: list, field_mapping: 
 
     for key in dto_instance.__fields__:  # type: ignore
         if key not in MyDTO.dto_field_mapping:
-            assert model_instance.__getattribute__(key) == dto_instance.__getattribute__(key)  # type: ignore
+            attribute_value = _get_attribute_value(model_instance, key)
+            assert attribute_value == dto_instance.__getattribute__(key)  # type: ignore
         else:
             original_key = MyDTO.dto_field_mapping[key]
-            assert model_instance.__getattribute__(original_key) == dto_instance.__getattribute__(key)  # type: ignore
+            attribute_value = _get_attribute_value(model_instance, original_key)
+            assert attribute_value == dto_instance.__getattribute__(key)  # type: ignore
 
 
 @pytest.mark.skipif(sys.version_info < (3, 9), reason="dataclasses behave differently in lower versions")
@@ -46,6 +58,7 @@ def test_conversion_to_model_instance(model: Any, exclude: list, field_mapping: 
     [
         [Person, ["id"], {"complex": "ultra"}, []],
         [VanillaDataClassPerson, ["id"], {"complex": "ultra"}, []],
+        [TypedDictPerson, ["id"], {"complex": "ultra"}, []],
         [Pet, ["age"], {"species": "kind"}, [SQLAlchemyPlugin()]],
     ],
 )
@@ -54,7 +67,7 @@ def test_conversion_from_model_instance(
 ) -> None:
     DTO = DTOFactory(plugins=plugins)("MyDTO", model, exclude=exclude, field_mapping=field_mapping)
 
-    if issubclass(model, (Person, VanillaDataClassPerson)):
+    if issubclass(model, (Person, VanillaDataClassPerson)) or is_typeddict(model):
         model_instance = model(
             first_name="moishe",
             last_name="zuchmir",
@@ -74,10 +87,10 @@ def test_conversion_from_model_instance(
     dto_instance = DTO.from_model_instance(model_instance=model_instance)
     for key in dto_instance.__fields__:
         if key not in DTO.dto_field_mapping:
-            assert model_instance.__getattribute__(key) == dto_instance.__getattribute__(key)
+            assert _get_attribute_value(model_instance, key) == _get_attribute_value(dto_instance, key)
         else:
             original_key = DTO.dto_field_mapping[key]
-            assert model_instance.__getattribute__(original_key) == dto_instance.__getattribute__(key)
+            assert _get_attribute_value(model_instance, original_key) == _get_attribute_value(dto_instance, key)
 
 
 async def test_async_conversion_from_model_instance(scaffold_tortoise: Callable, anyio_backend: str) -> None: